[Feat] 쿠폰 및 주문 취소 관련 수정

+++ backend/src/main/java/com/example/backend/coupon/controller/CouponController.java
@@ -9,6 +9,7 @@ import com.example.backend.coupon.model.dto.request.UserCouponCreateRequestDto;
 import com.example.backend.coupon.service.CouponService;
 import com.example.backend.global.response.BaseResponse;
 import com.example.backend.global.response.BaseResponseServiceImpl;
+import com.example.backend.global.response.responseStatus.CommonResponseStatus;
 import com.example.backend.global.response.responseStatus.CouponResponseStatus;
 import com.example.backend.user.model.User;
 import io.swagger.v3.oas.annotations.Operation;
@@ -27,6 +28,26 @@ import java.util.List;
 @RequestMapping("/coupon")
 public class CouponController {
     private final CouponService couponService;
+
+    @Operation( summary="이벤트 목록 조회", description="현재 진행 중인 이벤트 목록을 봅니다.")
+    @GetMapping("/events")
+    public BaseResponse<CouponListResponseDto> getEventList() {
+        return new BaseResponseServiceImpl().getSuccessResponse(couponService.getEventList(), CouponResponseStatus.SUCCESS);
+    }
+
+    @Operation( summary= "이벤트 쿠폰 발급", description="이벤트에서 쿠폰을 발행하고 이벤트 쿠폰 재고를 차감합니다")
+    @GetMapping("/events/{idx}")
+    public BaseResponse<Object> issueEventCoupon(@AuthenticationPrincipal User user, @PathVariable Long idx) {
+        // 중복 발행 방지
+        if (user.getUserCoupons() != null && user.getUserCoupons().stream().anyMatch(coupon -> coupon.getCoupon().getCouponIdx().equals(idx))) {
+            return new BaseResponseServiceImpl().getFailureResponse(CouponResponseStatus.DUPLICATED_EVENT_COUPON);
+        }
+        Boolean result = couponService.issueEventCoupon(user,idx);
+        if (!result) { // 쿠폰 재고 소진됨
+            return new BaseResponseServiceImpl().getFailureResponse(CouponResponseStatus.OUT_OF_COUPON_STOCK);
+        }
+        return new BaseResponseServiceImpl().getSuccessResponse(CouponResponseStatus.SUCCESS);
+    }
 
     @Operation(summary = "쿠폰 목록 조회", description = "전체 발급된 쿠폰 목록을 전부 조회합니다.")
     @GetMapping
@@ -52,25 +73,14 @@ public class CouponController {
         return new BaseResponseServiceImpl().getSuccessResponse(result, CouponResponseStatus.SUCCESS);
     }
 
-    @Operation( summary= "이벤트 쿠폰 발급", description="이벤트에서 쿠폰을 발행하고 이벤트 쿠폰 재고를 차감합니다")
-    @GetMapping("/events/{idx}")
-    public BaseResponse<Object> issueEventCoupon(@AuthenticationPrincipal User user, @PathVariable Long idx) {
-        // 중복 발행 방지
-        if (user.getUserCoupons().stream().anyMatch(coupon -> coupon.getCoupon().getCouponIdx().equals(idx))) {
-            return new BaseResponseServiceImpl().getFailureResponse(CouponResponseStatus.DUPLICATED_EVENT_COUPON);
-        }
-        Boolean result = couponService.issueEventCoupon(user,idx);
-        if (!result) { // 쿠폰 재고 소진됨
-            return new BaseResponseServiceImpl().getFailureResponse(CouponResponseStatus.OUT_OF_COUPON_STOCK);
-        }
-        return new BaseResponseServiceImpl().getSuccessResponse(CouponResponseStatus.SUCCESS);
-    }
-
     // ---------------- 관리자 ---------------------
 
     @Operation(summary="사용자별 쿠폰 발급", description="개별 사용자마다 수동 쿠폰 발급")
     @PostMapping("/issue")
-    public BaseResponse<String> issueCoupon(@RequestBody UserCouponCreateRequestDto request) {
+    public BaseResponse<String> issueCoupon(@AuthenticationPrincipal User user, @RequestBody UserCouponCreateRequestDto request) {
+        if (user == null || !user.getIsAdmin()) {
+            return new BaseResponseServiceImpl().getFailureResponse(CommonResponseStatus.BAD_REQUEST);
+        }
         Long couponIdx = couponService.CreateCouponForUser(request);
         log.info("issue coupon {}", couponIdx);
         // TODO: 사용자 알림 생성
@@ -80,7 +90,10 @@ public class CouponController {
 
     @Operation(summary = "전체 쿠폰 발급", description = "전체에게 쿠폰 발급.")
     @PostMapping("/issueall")
-    public BaseResponse<List<Long>> issueCouponsToAll(@RequestBody AllCouponCreateRequestDto request) {
+    public BaseResponse<List<Long>> issueCouponsToAll(@AuthenticationPrincipal User user, @RequestBody AllCouponCreateRequestDto request) {
+        if (user == null || !user.getIsAdmin()) {
+            return new BaseResponseServiceImpl().getFailureResponse(CommonResponseStatus.BAD_REQUEST);
+        }
         List<Long> coupons = couponService.CreateCouponForAll(request);
         log.info("issued {} coupons", coupons.size());
         return new BaseResponseServiceImpl().getSuccessResponse(coupons, CouponResponseStatus.SUCCESS);
@@ -89,14 +102,20 @@ public class CouponController {
 
     @Operation(summary = "쿠폰 수정", description= "쿠폰 내용 수정")
     @PutMapping("/update/{idx}")
-    public BaseResponse<String> updateCoupon(@PathVariable Long idx, @RequestBody UserCouponCreateRequestDto request) {
+    public BaseResponse<String> updateCoupon(@AuthenticationPrincipal User user, @PathVariable Long idx, @RequestBody UserCouponCreateRequestDto request) {
+        if (user == null || !user.getIsAdmin()) {
+            return new BaseResponseServiceImpl().getFailureResponse(CommonResponseStatus.BAD_REQUEST);
+        }
         couponService.updateCoupon(idx, request);
         return new BaseResponseServiceImpl().getSuccessResponse("수정 성공", CouponResponseStatus.SUCCESS);
     }
 
     @Operation(summary = "제한적 쿠폰 삭제", description= "단 한 명도 사용하지 않은 쿠폰 삭제")
     @DeleteMapping("/delete")
-    public BaseResponse<String> deleteCoupon(@RequestParam Long idx) {
+    public BaseResponse<String> deleteCoupon(@AuthenticationPrincipal User user, @RequestParam Long idx) {
+        if (user == null || !user.getIsAdmin()) {
+            return new BaseResponseServiceImpl().getFailureResponse(CommonResponseStatus.BAD_REQUEST);
+        }
         Boolean result = couponService.deleteCoupon(idx);
         if (result) {
             return new BaseResponseServiceImpl().getSuccessResponse("성공", CouponResponseStatus.SUCCESS);
@@ -107,7 +126,10 @@ public class CouponController {
 
     @Operation(summary = "쿠폰 검색", description = "키워드가 포함된 쿠폰 제목에 따라 쿠폰을 검색합니다.")
     @GetMapping("/search")
-    public BaseResponse<CouponListResponseDto> searchCoupon(@RequestParam String keyword) {
+    public BaseResponse<CouponListResponseDto> searchCoupon(@AuthenticationPrincipal User user, @RequestParam String keyword) {
+        if (user == null || !user.getIsAdmin()) {
+            return new BaseResponseServiceImpl().getFailureResponse(CommonResponseStatus.BAD_REQUEST);
+        }
         if (keyword == null) {
             keyword = "";
         }
@@ -118,21 +140,30 @@ public class CouponController {
 
     @Operation(summary = "선착순 쿠폰 발급 이벤트 등록", description = "선착순 쿠폰 발급 이벤트를 등록합니다")
     @PostMapping("/events")
-    public BaseResponse<Object> registerEvents(@RequestBody EventCouponCreateRequestDto request) {
+    public BaseResponse<Object> registerEvents(@AuthenticationPrincipal User user, @RequestBody EventCouponCreateRequestDto request) {
+        if (user == null || !user.getIsAdmin()) {
+            return new BaseResponseServiceImpl().getFailureResponse(CommonResponseStatus.BAD_REQUEST);
+        }
         couponService.createEvent(request);
         return new BaseResponseServiceImpl().getSuccessResponse(CouponResponseStatus.SUCCESS);
     }
 
     @Operation( summary= "선착순 쿠폰 발급 이벤트 수정" , description="예약된 이벤트를 수정합니다")
     @PutMapping("/events/{eventIdx}")
-    public BaseResponse<Object> updateEvents(@RequestBody EventCouponCreateRequestDto request, @PathVariable Long eventIdx) {
+    public BaseResponse<Object> updateEvents(@AuthenticationPrincipal User user, @RequestBody EventCouponCreateRequestDto request, @PathVariable Long eventIdx) {
+        if (user == null || !user.getIsAdmin()) {
+            return new BaseResponseServiceImpl().getFailureResponse(CommonResponseStatus.BAD_REQUEST);
+        }
         couponService.updateEvent(eventIdx, request);
         return new BaseResponseServiceImpl().getSuccessResponse(CouponResponseStatus.SUCCESS);
     }
 
     @Operation( summary= "선착순 쿠폰 발급 이벤트 삭제" , description="발행한 이벤트를 삭제합니다. 이 쿠폰은 사용한 것은 삭제되지 않으므로 주의해야 합니다.")
     @DeleteMapping("/events")
-    public BaseResponse<Object> deleteEvents(@RequestParam Long idx) {
+    public BaseResponse<Object> deleteEvents(@AuthenticationPrincipal User user, @RequestParam Long idx) {
+        if (user == null || !user.getIsAdmin()) {
+            return new BaseResponseServiceImpl().getFailureResponse(CommonResponseStatus.BAD_REQUEST);
+        }
         couponService.forceDeleteEvent(idx);
         return new BaseResponseServiceImpl().getSuccessResponse(CouponResponseStatus.SUCCESS);
     }
